requirements/dev.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Pydantic==0.12.0
Flask-WTF==1.2.1
alembic==1.13.1
# psycopg2==2.9.9
pyjwt==2.8.0
jsonschema==4.21.0
python-dotenv==1.0.1
tb-rest-client==3.6.2
flask-marshmallow==1.1.0
marshmallow-sqlalchemy==0.30.0
